Student-Projects/group17/requirements.txt
python-telegram-bot==21.10
groq==0.16.0
python-dotenv==1.0.1